@@ -31,40 +28,51 @@
     "@backstage/plugin-catalog-graph": "workspace:^",
     "@backstage/plugin-catalog-import": "workspace:^",
     "@backstage/plugin-catalog-react": "workspace:^",
-    "@backstage/plugin-github-actions": "workspace:^",
-    "@backstage/plugin-org": "workspace:^",
+    "@backstage/plugin-catalog-unprocessed-entities": "workspace:^",
+    "@backstage/plugin-home": "workspace:^",
     "@backstage/plugin-permission-react": "workspace:^",
     "@backstage/plugin-scaffolder": "workspace:^",
     "@backstage/plugin-scaffolder-react": "workspace:^",
     "@backstage/plugin-search": "workspace:^",
+    "@backstage/plugin-search-common": "workspace:^",
     "@backstage/plugin-search-react": "workspace:^",
-    "@backstage/plugin-tech-radar": "workspace:^",
-    "@backstage/plugin-techdocs": "workspace:^",
-    "@backstage/plugin-techdocs-module-addons-contrib": "workspace:^",
-    "@backstage/plugin-techdocs-react": "workspace:^",
     "@backstage/plugin-user-settings": "workspace:^",
     "@backstage/theme": "workspace:^",
     "@material-ui/core": "^4.12.2",
     "@material-ui/icons": "^4.9.1",
+    "@material-ui/lab": "4.0.0-alpha.61",
+    "@octokit/rest": "^19.0.3",
+    "@vitejs/plugin-react": "^4.0.4",
     "history": "^5.0.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.2",
+    "react-dom": "^18.0.2",
     "react-router": "^6.3.0",
     "react-router-dom": "^6.3.0",
-    "react-use": "^17.2.4"
+    "react-use": "^17.2.4",
+    "vite": "^4.4.9",
+    "vite-plugin-html": "^3.2.0",
+    "vite-plugin-node-polyfills": "^0.16.0",
+    "zen-observable": "^0.10.0"
   },
   "devDependencies": {
     "@backstage/test-utils": "workspace:^",
-    "@testing-library/dom": "^8.0.0",
-    "@testing-library/jest-dom": "^5.10.1",
-    "@testing-library/react": "^12.1.3",
+    "@playwright/test": "^1.32.3",
+    "@testing-library/dom": "^9.0.0",
+    "@testing-library/jest-dom": "^6.0.0",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.0.0",
-    "@types/node": "^16.11.26",
+    "@types/jquery": "^3.3.34",
+    "@types/react": "*",
     "@types/react-dom": "*",
-    "cross-env": "^7.0.0",
-    "cypress": "^10.0.0",
-    "eslint-plugin-cypress": "^2.10.3",
-    "start-server-and-test": "^1.10.11"
+    "@types/zen-observable": "^0.8.0",
+    "cross-env": "^7.0.0"
+  },
+  "scripts": {
+    "start": "backstage-cli package start",
+    "build": "backstage-cli package build",
+    "clean": "backstage-cli package clean",
+    "test": "backstage-cli package test",
+    "lint": "backstage-cli package lint"
   },
   "browserslist": {
     "production": [
